requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
base58==2.1.1
beartype==0.20.2
beautifulsoup4==4.13.3
bit==0.8.0
bitcoin==1.1.42
bitcoinlib==0.7.2
boolean.py==4.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
coincurve==20.0.0
colorama==0.4.6
contourpy==1.3.1
cryptography==41.0.3
cycler==0.12.1
EbookLib==0.18
ecdsa==0.19.0
fastapi==0.128.5
fonttools==4.56.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
license-expression==30.4.1
llvmlite==0.44.0
lxml==5.3.1
matplotlib==3.10.1
memory-profiler==0.61.0
numba==0.61.0
numpy==2.1.3
oauthlib==3.3.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
ply==3.11
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyGithub==2.8.1
PyJWT==2.11.0
PyMuPDF==1.25.4
PyNaCl==1.6.2
pyOpenSSL==23.2.0
pyparsing==3.2.3
python-bitcoinlib @ git+https://github.com/petertodd/python-bitcoinlib.git@06e7289ee268167953c38b40ab16ce4269cae583
python-bitcoinrpc==1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.3
redis==7.1.0
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
setuptools==75.8.2
six==1.17.0
soupsieve==2.6
spdx-tools==0.8.3
spotipy==2.25.2
SQLAlchemy==2.0.38
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
uritools==4.0.3
urllib3==2.3.0
uvicorn==0.40.0
wheel==0.45.1
xmltodict==0.14.2
